refactor(compositor): extract screenshot flow and reduce startup churn
- move screenshot session and capture handling out of `compositor/root.rs` into a dedicated compositor module so the flow is easier to maintain without changing behavior
- prewarm visible window offscreen caches so the first collapse-to-node animation can reuse a ready snapshot, and quiet noisy tty vblank diagnostics so persistent backend issues stand out in debug logs
