+++ crates/halley-wl/src/backend/vblank_throttle.rs
@@ -20,6 +20,11 @@ pub(crate) struct VBlankThrottle {
     max_jitter: Duration,
     min_interval: Option<Duration>,
     max_interval: Duration,
+    window_throttled_count: u64,
+    window_estimated_missed_vblanks: u64,
+    window_max_jitter: Duration,
+    window_min_interval: Option<Duration>,
+    window_max_interval: Duration,
     last_report_at: Option<Instant>,
 }
 
@@ -38,8 +43,50 @@ impl VBlankThrottle {
             max_jitter: Duration::ZERO,
             min_interval: None,
             max_interval: Duration::ZERO,
+            window_throttled_count: 0,
+            window_estimated_missed_vblanks: 0,
+            window_max_jitter: Duration::ZERO,
+            window_min_interval: None,
+            window_max_interval: Duration::ZERO,
             last_report_at: None,
         }
+    }
+
+    fn maybe_report_metrics(&mut self, timestamp: Instant) {
+        const REPORT_INTERVAL: Duration = Duration::from_secs(30);
+        const REPORT_JITTER_THRESHOLD: Duration = Duration::from_millis(250);
+
+        let should_report = match self.last_report_at {
+            Some(last) => timestamp.saturating_duration_since(last) >= REPORT_INTERVAL,
+            None => self.samples >= 240,
+        };
+        if !should_report {
+            return;
+        }
+
+        self.last_report_at = Some(timestamp);
+
+        let interesting = self.window_throttled_count > 0
+            || self.window_estimated_missed_vblanks > 0
+            || self.window_max_jitter >= REPORT_JITTER_THRESHOLD;
+        if interesting {
+            debug!(
+                "vblank pacing [{}]: throttled={} est_missed={} avg={:?} min={:?} max={:?} max_jitter={:?}",
+                self.output_name,
+                self.window_throttled_count,
+                self.window_estimated_missed_vblanks,
+                self.smoothed_interval.unwrap_or(Duration::ZERO),
+                self.window_min_interval.unwrap_or(Duration::ZERO),
+                self.window_max_interval,
+                self.window_max_jitter,
+            );
+        }
+
+        self.window_throttled_count = 0;
+        self.window_estimated_missed_vblanks = 0;
+        self.window_max_jitter = Duration::ZERO;
+        self.window_min_interval = None;
+        self.window_max_interval = Duration::ZERO;
     }
 
     fn update_metrics(
@@ -54,6 +101,11 @@ impl VBlankThrottle {
             None => passed,
         });
         self.max_interval = self.max_interval.max(passed);
+        self.window_min_interval = Some(match self.window_min_interval {
+            Some(current) => current.min(passed),
+            None => passed,
+        });
+        self.window_max_interval = self.window_max_interval.max(passed);
 
         let smoothed = match self.smoothed_interval {
             Some(current) => {
@@ -74,35 +126,21 @@ impl VBlankThrottle {
                 refresh - passed
             };
             self.max_jitter = self.max_jitter.max(jitter);
+            self.window_max_jitter = self.window_max_jitter.max(jitter);
 
             let refresh_ns = refresh.as_nanos();
             let passed_ns = passed.as_nanos();
             if refresh_ns > 0 && passed_ns > refresh_ns + (refresh_ns / 2) {
                 let missed = (passed_ns / refresh_ns).saturating_sub(1);
-                self.estimated_missed_vblanks = self
-                    .estimated_missed_vblanks
-                    .saturating_add(missed.min(u128::from(u64::MAX)) as u64);
+                let missed = missed.min(u128::from(u64::MAX)) as u64;
+                self.estimated_missed_vblanks =
+                    self.estimated_missed_vblanks.saturating_add(missed);
+                self.window_estimated_missed_vblanks =
+                    self.window_estimated_missed_vblanks.saturating_add(missed);
             }
         }
 
-        let should_report = match self.last_report_at {
-            Some(last) => timestamp.saturating_duration_since(last) >= Duration::from_secs(5),
-            None => self.samples >= 120,
-        };
-        if should_report {
-            self.last_report_at = Some(timestamp);
-            debug!(
-                "vblank pacing [{}]: samples={} throttled={} est_missed={} avg={:?} min={:?} max={:?} max_jitter={:?}",
-                self.output_name,
-                self.samples,
-                self.throttled_count,
-                self.estimated_missed_vblanks,
-                self.smoothed_interval.unwrap_or(Duration::ZERO),
-                self.min_interval.unwrap_or(Duration::ZERO),
-                self.max_interval,
-                self.max_jitter,
-            );
-        }
+        self.maybe_report_metrics(timestamp);
     }
 
     pub(crate) fn throttle(
@@ -131,6 +169,7 @@ impl VBlankThrottle {
                     }
 
                     self.throttled_count = self.throttled_count.saturating_add(1);
+                    self.window_throttled_count = self.window_throttled_count.saturating_add(1);
                     let remaining = refresh.saturating_sub(passed);
                     let token = self
                         .event_loop
